requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
django-rest-swagger
django-rest-authtoken
Pillow
django-jazzmin
django-environ
environ
psycopg2-binary
psycopg2